pyproject.toml
[project]
name = "genome-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.13",
    "mwclient>=0.11.0",
    "mwparserfromhell>=0.7.0",
    "polars>=1.31.0",
    "streamlit>=1.46.1",
]

[dependency-groups]
dev = [
    "google-genai>=1.23.0",
    "marimo>=0.14.8",
]
